mv rethnet_eth::remote filter stuff to its own mod
